Refactor project structure and convert scripts to modules
