+++ public/index.html
@@ -9,7 +9,7 @@
     <script type="module" src="https://ajax.googleapis.com/ajax/libs/model-viewer/3.5.0/model-viewer.min.js"></script>
     <script src="https://unpkg.com/@ffmpeg/ffmpeg@0.11.0/dist/ffmpeg.min.js"></script>
     <link rel="stylesheet" href="https://unpkg.com/7.css">
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="css/styles.css">
     <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" rel="stylesheet">
     <link rel="stylesheet" href="https://code.jquery.com/ui/1.13.2/themes/base/jquery-ui.css">
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.4/jquery.min.js"></script>
@@ -113,12 +113,12 @@
                 <p><b>Not affiliated with ppy Pty Ltd or Chavafei.<br>"osu!" is a trademark of ppy Pty Ltd.</b></p>
                 <p>License: MIT</p>
                 <div class="field-row">
-                    <button id="link1" onclick="link1click()">GitHub</button>
-                    <button id="link2" onclick="link2click()">Sinker Sound Discord</button>
+                    <button id="github-btn">GitHub</button>
+                    <button id="discord-btn">Sinker Sound Discord</button>
                 </div>
                 <p>UI made with <a href="https://khang-nd.github.io/7.css/">7.css</a> library by Khang (khang-nd).</p>
                 </p>
-                <button id="link1" onclick="showlogs()">Show log</button>
+                <button id="log-btn">Show log</button>
                 <button id="dvd-mode-btn">DVD mode</button>
                 <button id="gravity-btn">Gravity on</button>
             </article>
@@ -147,7 +147,7 @@
             and cannot convert maps made for osu!catch, taiko, or mania.
         </div>
         <footer style="text-align: right">
-            <button onclick="copylog()">Copy logs</button>
+            <button id="copylog-btn">Copy logs</button>
             <button aria-label="Close">OK</button>
         </footer>
     </div>
@@ -163,7 +163,7 @@
             Unfortunately, you cannot close the logo.
         </div>
         <footer style="text-align: right">
-            <button onclick="learnmore()">Learn more</button>
+            <button id="learnmore-btn">Learn more</button>
             <button aria-label="Close">OK</button>
         </footer>
     </div>
@@ -202,9 +202,4 @@
     </div>
 </div>
 <div id="taskbar"></div>
-<script src="win7.js"></script>
-<script src="converter.js"></script>
-<script src="ui.js"></script>
-<!-- here comes the fun stuff -->
-<script src="dvd.js"></script>
-<script src="gravity.js"></script>
+<script type="module" src="js/main.js"></script>
