+++ public/js/win7.js
@@ -1,3 +1,4 @@
+export function initTabs() {
 // Tabs
 function tabHandler(e, tabButtons) {
   e.preventDefault();
@@ -83,3 +84,4 @@ document
       }
     });
   });
+}
